updated initiative ordering, was having difficulties keeping it up to date while adding or removing entities, going to keep the initiative manager in the session settings window, however upon starting combat you will decide the order of initiative and it will go from there. If there has already been an initiative order decided then the start combat button will only flip the 'mode' column of the game table to be 'combat' or 'noncombat'.
Keeping the initiative manager will allow for changes made in the middle
of combat to not mess with the initiative. Adding an entity will place
it at the end of the move order by default but the initiative manager
will allow you to adjust its place in the order.

Deciding on how best to account how entities being removed. I dont want
to bother with redefining the order but I will likely have to. I want to
keep it in the same format with the trailing comma so we will see what I
have to do to maintain that. if its too much of a pain its not worth it.